// libs/remix-ui/home-tab/src/lib/components/types/carouselTypes.ts
export interface UpdateInfo {
  badge: string
  title: string
  description: string
  descriptionList?: string[]
  icon: string
  action: {
    type: 'link' | 'methodCall'
    label: string
    url?: string
    pluginName?: string
    pluginMethod?: string,
    pluginArgs?: (string | number | boolean | object | null)[]
  },
  theme: string
}